requirements.txt
fastapi
uvicorn
python-multipart
pymongo
pydantic
pydantic-settings
sqlalchemy
passlib[argon2]
python-jose[cryptography]
argon2-cffi
